+++ packages/client/src/socketio.js
@@ -1,0 +1,1 @@
+module.exports = require('@feathersjs/socketio-client');